backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
mistralai==0.0.12
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
edge-tts==6.1.9
gunicorn==21.2.0
